refactor arrow loop image as SVG, update link styles
