+++ components/LandingPage/Hero/Hero.tsx
@@ -1,4 +1,4 @@
-import { Image } from 'components/Image/Image';
+import Image from 'components/Image';
 import styles from './Hero.module.scss';
 import { ImageTypes } from '../../../types/Base';
 
@@ -35,6 +35,7 @@ export const Hero = ({ image, children }: HeroProps) => {
                     src={image.url}
                     alt={image.description}
                     className={styles.image}
+                    priority
                 />
                 <div className={styles.textContainer}>
                     <h1 className={styles.title}>
