+++ components/LandingPage/LifeAsATorchboxer/LifeAsATorchboxer.module.scss
@@ -72,7 +72,7 @@
 
     @include media-query(large) {
         order: 1;
-        margin: auto $grid * 2 $grid * 2 auto;
+        margin: auto $grid * 2 0 auto;
         transform: rotate(0deg);
     }
 }
