+++ components/LifeAtTorchbox/Hero/Hero.tsx
@@ -1,5 +1,5 @@
 import { useRef, useState } from 'react';
-import { ArrowLoopRight } from 'components/Icons/ArrowLoopRight';
+import { ArrowLoopRight } from 'components/SVG/ArrowLoop/ArrowLoopRight';
 import Image from 'components/Image';
 import { ImageTypes, VideoTypes } from '../../../types/Base';
 import { VideoPlayPauseButton } from './VideoPlayPauseButton';
@@ -61,6 +61,7 @@ export const Hero = ({ image, video, subtitle, children }: HeroProps) => {
                     src={image.url}
                     alt={image.description}
                     className={styles.mobileImage}
+                    priority
                 />
                 {video ? (
                     <>
@@ -88,6 +89,7 @@ export const Hero = ({ image, video, subtitle, children }: HeroProps) => {
                         src={image.url}
                         alt={image.description}
                         className={styles.desktopImage}
+                        priority
                     />
                 )}
                 <div className={styles.textContainer}>
