+++ components/LifeAtTorchbox/ValuesCarousel/ValuesCarousel.module.scss
@@ -41,7 +41,7 @@
         display: block;
         position: relative;
         order: 1;
-        margin: auto $grid * 3 $grid * 2 auto;
+        margin: auto $grid * 2 $grid * 2 auto;
         bottom: -60px;
     }
 }
