+++ styles/_utilities.scss
@@ -1,7 +1,7 @@
 .underline-link {
-    @include font-size(s);
+    font-size: 22px;
     color: var(--color--white);
-    font-weight: $weight--bold;
+    font-weight: $weight--extrabold;
     border-bottom: 0;
     overflow: visible;
     position: relative;
@@ -16,10 +16,6 @@
         &:focus {
             outline: 4px solid var(--color--coral) transparent;
         }
-    }
-
-    &--extrabold {
-        font-weight: $weight--extrabold;
     }
 
     &--indigo {
@@ -51,7 +47,7 @@
         position: absolute;
         width: 65px;
         height: 6px;
-        bottom: -7px;
+        bottom: -3px;
         left: 0;
         background-color: var(--color--coral);
         transition: width $transition-quick;
